requirements.txt
aiohttp==3.8.4
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
bidict==0.22.1
botocore==1.29.133
cached-property==1.5.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.2
Deprecated==1.2.13
docker==3.7.2
docker-pycreds==0.4.0
docutils==0.18.1
fastapi==0.92.0
google-auth==2.18.0
greenlet==2.0.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
install==1.3.5
jdcal==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
jsonify==0.5
kombu==5.2.4
kubernetes==26.1.0
ldap3==2.9
lml==0.0.9
logger==1.4
login==0.0.6
loguru==0.7.0
lxml==4.9.1
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.0
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
pip-review==1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.8.1
pydantic==1.10.5
pyexcel==0.5.13
pyexcel-io==0.5.17
pyexcel-webio==0.1.4
Pygments==2.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.3.1
python-multipart==0.0.5
python-socketio==5.5.2
pytz==2022.1
PyYAML==6.0
redis==4.2.2
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
slackclient==1.3.2
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utc==0.14.0
sqlparse==0.4.2
starlette==0.25.0
tencentcloud-sdk-python==3.0.543
termcolor==1.1.0
texttable==1.6.1
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.5.0
tzlocal==2.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.14.0
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.54.0
websockets==11.0.2
Werkzeug==2.2.3
wrapt==1.14.0
xlrd==2.0.1
yarl==1.6.3
zipp==3.15.0
zope.interface==5.4.0
